ydb_federated_topic: Move fds_mock to separate directory
... to use it in ydb_persqueue_over_topic tests
