+++ ydb/public/sdk/cpp/client/ydb_federated_topic/ut/fds_mock/ya.make
@@ -1,0 +1,8 @@
+LIBRARY()
+
+SRCS(
+    fds_mock.h
+)
+
+END()
+
