+++ ydb/public/sdk/cpp/client/ydb_federated_topic/ut/ya.make
@@ -27,6 +27,7 @@ PEERDIR(
 
     ydb/public/sdk/cpp/client/ydb_federated_topic
     ydb/public/sdk/cpp/client/ydb_federated_topic/impl
+    ydb/public/sdk/cpp/client/ydb_federated_topic/ut/fds_mock
 )
 
 YQL_LAST_ABI_VERSION()
